backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
transformers==4.37.2
torch==2.1.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
pypdf==3.17.4
huggingface-hub==0.20.3